Componentize EventTime (now it's shared between PostsPage and PostsItemMeta)
Factor EventTime out into a component, without (yet) modifying it to
handle timezones.
