+++ packages/lesswrong/components/posts/PostsPage.jsx
@@ -99,12 +99,6 @@ const styles = theme => ({
       lineHeight: 1.25,
       fontWeight: 600,
     },
-    eventTimeStart: {
-      display: "inline-block",
-    },
-    eventTimeEnd: {
-      display: "inline-block",
-    },
     
     eventLocation: {
       fontSize: "14px",
@@ -212,52 +206,12 @@ class PostsPage extends Component {
     const { classes } = this.props
     if (post.isEvent) {
       return <div className={classes.eventTimes}>
-        {this.renderEventTimes(post.startTime, post.endTime)}
+        <Components.EventTime post={post}/>
       </div>
     } else {
       return <div className={classes.subtitle}>
         {moment(post.postedAt).format('MMM D, YYYY')}
       </div>
-    }
-  }
-  
-  renderEventTimes = (start, end) => {
-    const classes = this.props.classes;
-    const timeFormat = 'h:mm A';
-    const dateFormat = 'MMMM Do YY, '+timeFormat
-    const calendarFormat = {sameElse : dateFormat}
-    
-    // Neither start nor end time specified
-    if (!start && !end) {
-      return "TBD";
-    }
-    // Start time specified, end time missing. Use
-    // moment.calendar, which has a bunch of its own special
-    // cases like "tomorrow".
-    // (Or vise versa. Specifying end time without specifying start time makes
-    // less sense, but users can enter silly things.)
-    else if (!start || !end) {
-      const eventTime = start ? start : end;
-      return moment(eventTime).calendar({}, calendarFormat)
-    }
-    // Both start end end time specified
-    else {
-      // If the start and end time are on the same date, render it like:
-      //   January 15 13:00-15:00
-      // If they're on different dates, render it like:
-      //   January 15 19:00 to January 16 12:00
-      if (moment(start).format("YYYY-MM-DD") === moment(end).format("YYYY-MM-DD")) {
-        return moment(start).format(dateFormat) + '-' + moment(end).format(timeFormat);
-      } else {
-        return (<span>
-          <span className={classes.eventTimeStart}>
-            From: {moment(start).calendar({}, calendarFormat)}
-          </span>
-          <span className={classes.eventTimeEnd}>
-            To: {moment(end).calendar({}, calendarFormat)}
-          </span>
-        </span>);
-      }
     }
   }
